Extract ShippingContext from CheckoutContext
