+++ src/Sylius/Behat/Resources/config/suites/ui/checkout/checkout.yml
@@ -48,6 +48,7 @@ default:
                 - sylius.behat.context.ui.channel
                 - sylius.behat.context.ui.checkout
                 - sylius.behat.context.ui.checkout.addressing
+                - sylius.behat.context.ui.checkout.shipping
                 - sylius.behat.context.ui.checkout.thank_you
                 - sylius.behat.context.ui.shop.address_book
                 - sylius.behat.context.ui.shop.cart
